pyproject.toml
[project]
name = "cirobotix"
version = "0.1.0"
description = "AI-driven CI/CD automation framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.7",
  "PyYAML>=6.0",
  "requests>=2.31",
  "rich>=13.7",
  "pytest>=8.4.2",

]

[project.scripts]
cirobotix = "cirobotix.cli:app"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
